Horizontal and vertical hex grids do their own grid offsets
With scaling in play, we can't rely on the width and height being larger or smaller than the other.
